se realiza la funcion de crear post editar y eliminar solo por la persona que lo publico, tambien se arregla los estilos de las vistas feed y post. se adecua que solo se pueda realizar un like por usuario y se le dio funcionalidad al btn de feed para poder volver al feed general

+++ src/components/feed.js
@@ -23,6 +23,9 @@ async function feed(navigateTo) {
     let postsContainer = '';
     if (userPosts !== undefined) {
       userPosts.forEach((post, index) => {
+        console.log(post.likes);
+        const likesNumber = post.likes.length;
+        const isLikedByUser = post.likes.some(like => like.user === user.uid);
         postsContainer += `
         <div class="card" id="card-${post.id}">
           <div class="container-creator">
@@ -33,13 +36,15 @@ async function feed(navigateTo) {
             <p id="description-${post.id}">${post.description}</p>
           </div>
           <div class="container-heart">
-            <h3 id="likes-number-${post.id}">${post.likes}</h3>
-            <img id="like-img-${post.id}" data='{"user":"${user.uid}","post":"${post.id}"}' class="like-img" src="../components/images/corazon2.png" />
+            <h3 id="likes-number-${post.id}" class="${(isLikedByUser) ? '' : 'noLiked'}">${likesNumber}</h3>
+            <img id="like-img-${post.id}" data='{"user":"${user.uid}","post":"${post.id}", "isLikedByUser": "${isLikedByUser}"}' class="like-img" src="${(isLikedByUser) ? '../components/images/corazon2.png' : '../components/images/corazon.png'}" />
           </div>
-          <div class="container-admin-btns">
-            <img id="upd-img-${post.id}" data='{"user":"${user.uid}","post":"${post.id}"}' class="upd-img" src="../components/images/editar.png" />
-            <img id="del-img-${post.id}" data='{"user":"${user.uid}","post":"${post.id}"}' class="del-img" src="../components/images/cancelar.png" />
-          </div>
+          ${user.uid === post.createdBy ?
+            `<div class="container-admin-btns">
+               <img id="upd-img-${post.id}" data='{"user":"${user.uid}","post":"${post.id}"}' class="upd-img" src="../components/images/editar.png" />
+               <img id="del-img-${post.id}" data='{"user":"${user.uid}","post":"${post.id}"}' class="del-img" src="../components/images/cancelar.png" />
+             </div>` : ''
+          }
         </div>
         `;
       });
@@ -81,15 +86,24 @@ async function feed(navigateTo) {
     return footer;
   }
 
-  async function applySendingLikesOption(sectionFeed) {
+  async function applySendingLikesOption(sectionFeed, user) {
     const likeBtns = sectionFeed.querySelectorAll('.like-img');
     likeBtns.forEach((heart) => {
       heart.addEventListener('click', async () => {
         const likeData = JSON.parse(heart.getAttribute('data'));
-        const likesContainer = sectionFeed.querySelector(`#likes-number-${likeData.post}`);
-        const likesNumer = parseInt(likesContainer.textContent) + 1;
-        await updatePostLikes(likeData.post, likesNumer);
-        likesContainer.textContent = likesNumer;
+        console.log(likeData);
+        console.log(JSON.parse(likeData.isLikedByUser));
+        if(!JSON.parse(likeData.isLikedByUser)){
+          const likesContainer = sectionFeed.querySelector(`#likes-number-${likeData.post}`);
+          const likesNumer = parseInt(likesContainer.textContent) + 1;
+          await updatePostLikes(likeData.post, {user: user.uid});
+          likesContainer.textContent = likesNumer;
+          heart.setAttribute('data',JSON.stringify({...likeData, "isLikedByUser": "true"}));
+          heart.setAttribute("src", "../components/images/corazon2.png");
+          likesContainer.classList.remove("noLiked");
+        } else {
+          console.log("El usuario ya le habia dado like antes");
+        }
       });
     });
   }
@@ -98,10 +112,12 @@ async function feed(navigateTo) {
     const delBtns = sectionFeed.querySelectorAll('.del-img');
     delBtns.forEach((post) => {
       post.addEventListener('click', async () => {
-        const likeData = JSON.parse(post.getAttribute('data'));
-        await deletePost(likeData.post);
-        const card = sectionFeed.querySelector(`#card-${likeData.post}`);
-        card.remove();
+        if(confirm("Estas seguro que desea eliminar el post")){
+          const likeData = JSON.parse(post.getAttribute('data'));
+          await deletePost(likeData.post);
+          const card = sectionFeed.querySelector(`#card-${likeData.post}`);
+          card.remove();
+        }
       });
     });
   }
@@ -144,7 +160,7 @@ async function feed(navigateTo) {
     ${createFooter().outerHTML}
   `;
   sectionFeed.innerHTML = feedSection;
-  applySendingLikesOption(sectionFeed);
+  applySendingLikesOption(sectionFeed, user);
   applyDeletePostOption(sectionFeed);
   applyUpdatePostOption(sectionFeed);
   applyCreatePostFunction(sectionFeed);
